Fix signess warning.

+++ src/conference/session/media-session.cpp
@@ -1119,7 +1119,7 @@ int MediaSessionPrivate::selectRandomPort (int streamIndex, pair<int, int> portR
 	for (int nbTries = 0; nbTries < 100; nbTries++) {
 		bool alreadyUsed = false;
 		unsigned int rangeSize = static_cast<unsigned int>(portRange.second - portRange.first);
-		unsigned int randomInRangeSize = (ortp_random() % rangeSize) & ~0x1; /* Select an even number */
+		unsigned int randomInRangeSize = (ortp_random() % rangeSize) & (unsigned int)~0x1; /* Select an even number */
 		int triedPort = ((int)randomInRangeSize) + portRange.first;
 		/*If portRange.first is even, the triedPort will be even too. The one who configures a port range that starts with an odd number will
 		 * get odd RTP port numbers.*/
